:construction: Move extension method inside of the table view source
